Renaming atmosphere corrections files
