+++ include/tudat/astro/observation_models.h
@@ -32,7 +32,7 @@
 #include "observation_models/relativeAngularPositionObservationModel.h"
 #include "observation_models/corrections/lightTimeCorrection.h"
 #include "observation_models/corrections/firstOrderRelativisticCorrection.h"
-#include "observation_models/corrections/tabulatedMediaCorrection.h"
+#include "observation_models/corrections/atmosphereCorrection.h"
 #include "observation_models/corrections/solarCoronaCorrection.h"
 
 #endif // TUDAT_OBSERVATION_MODELS_H
